Finishing batch test fix

+++ tests/hip_vectoradd_cmakelists_batch.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 REPO_DIR="$(dirname "$(dirname "$(readlink -fm "$0")")")"
-sinfo
-sbatch --wait --time 3 ${REPO_DIR}/HIP/vectorAdd/hip_cmakelists_batch.sh
+if [[ `sinfo | grep unk | wc -l` == 1 ]]; then
+   echo "SLURM not configured properly -- SKIPPING"
+else
+   sbatch --wait ${REPO_DIR}/HIP/vectorAdd/hip_cmakelists_batch.sh
 
-grep PASSED! slurm-*.out
-ls
-rm  slurm-*.out
+   grep PASSED! slurm-*.out
+   rm  slurm-*.out
+fi
